Moved customEvent and eventTarget shim
